fix: move missing project warning to settings
